requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
auto-labeling-pipeline==0.1.23
backports-datetime-fromisoformat==2.0.3
billiard==4.2.1
boto3==1.38.46
botocore==1.38.46
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.8
Django==4.2.23
django-allauth==0.52.0
django-celery-results==2.4.0
django-cleanup==6.0.0
django-cors-headers==3.14.0
django-drf-filepond==0.5.0
django-filter==22.1
django-health-check==3.20.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework-xml==2.0.0
doccano==1.8.4
drf-yasg==1.21.10
environs==9.5.0
et_xmlfile==2.0.0
faiss-cpu==1.11.0
filetype==1.2.0
flower==1.2.0
furl==2.1.4
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
gunicorn==20.1.0
humanize==4.12.3
idna==3.10
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kombu==5.5.4
lml==0.2.0
MarkupSafe==3.0.2
marshmallow==4.0.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
openpyxl==3.1.5
orderedmultidict==1.0.1
packaging==25.0
pandas==1.5.3
prometheus_client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyexcel==0.7.3
pyexcel-io==0.6.7
pyexcel-xlsx==0.6.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.15.3
seqeval==1.2.2
shortuuid==1.0.13
six==1.17.0
SPARQLWrapper==2.0.0
SQLAlchemy==1.4.54
sqlparse==0.5.3
texttable==1.7.0
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
waitress==2.1.2
wcwidth==0.2.13
whitenoise==6.9.0
